+++ R/use.R
@@ -6,15 +6,15 @@ NULL
 #' Use CWB indexed corpora in R data packages by adding registry file to session
 #' registry.
 #'
-#' {pkg} is expected to be an installed data package that includes CWB indexed
-#' corpora. The \code{use}-function will add the registry files describing the
+#' `pkg` is expected to be an installed data package that includes CWB indexed
+#' corpora. The `use()`-function will add the registry files describing the
 #' corpus (or the corpora) to the session registry directory and adjust the path
 #' pointing to the data in the package.
 #'
 #' The registry files within the package are assumed to be in the subdirectory
-#' \code{./extdata/cwb/registry} of the installed package. The data directories
+#' './extdata/cwb/registry' of the installed package. The data directories
 #' for corpora are assumed to be in a subdirectory named after the corpus (lower
-#' case) in the package subdirectory \code{./extdata/cwb/indexed_corpora/}. When
+#' case) in the package subdirectory './extdata/cwb/indexed_corpora/'. When
 #' adding a corpus to the registry, templates for formatting fulltext output are
 #' reloaded.
 #' 
@@ -30,32 +30,48 @@ NULL
 #' @export use
 #' @rdname use
 #' @name use
+#' @return A `logical` value: `TRUE` if corpus has been loaded successfully, or
+#'   `FALSE`, if any kind of error occurred.
 #' @examples
 #' use("polmineR")
 #' corpus()
 #' @seealso To get the temporary registry directory, see \code{\link{registry}}.
 #' @importFrom RcppCWB cqp_reset_registry cl_load_corpus cqp_load_corpus
 #' @importFrom stringi stri_enc_mark
+#' @importFrom cli cli_alert_danger cli_alert_success
 use <- function(pkg, corpus, lib.loc = .libPaths(), tmp = FALSE, verbose = TRUE){
   
-  if (nchar(system.file(package = pkg)) == 0L)
-    stop("Could not find package specified. Please check for typos,",
-         "and/or whether it is installed for the R version you are using.")
-  
+  if (nchar(system.file(package = pkg)) == 0L){
+    if (verbose) cli_alert_danger(
+      paste0(
+        "Could not find package {.pkg {pkg}}. Please check for typos, ",
+        "and/or whether it is installed for the R version you are using."
+      )
+    )
+    return(FALSE)
+  }
+
   registry_dir <- system.file(
     "extdata", "cwb", "registry", package = pkg, lib.loc = lib.loc
   )
-  if (!dir.exists(registry_dir))
-    stop("pkg exists, but is not a standardized package - registry directory missing")
-  
+  if (!dir.exists(registry_dir)){
+    if (verbose) cli_alert_danger(
+      paste0(
+        "package {.pkg {pkg}} exists, but is not a standardized package - ",
+        "no registry directory found"
+      )
+    )
+    return(FALSE)
+  }
+    
   corpora <- list.files(registry_dir)
   if (!missing(corpus)){
     if (all(tolower(corpus) %in% corpora)){
       corpora <- tolower(corpus)
     } else {
-      stop("corpus provided is not included in package")
+      if (verbose) cli_alert_danger("corpus/corpora not included in package")
+      return(FALSE)
     }
-    
   }
 
   for (corpus in corpora){
@@ -71,20 +87,20 @@ use <- function(pkg, corpus, lib.loc = .libPaths(), tmp = FALSE, verbose = TRUE)
           "\\\\", "/", utils::shortPathName(corpus_data_srcdir)
         )
       } else {
-        # Copying files to a temporary directory that does not include non-ASCII characters
-        # may still be necessary on macOS machines
+        # Copying files to a temporary directory that does not include non-ASCII
+        # characters may still be necessary on macOS machines
         
         if (!dir.exists(corpus_data_targetdir))
           dir.create(corpus_data_targetdir)
         else
           file.remove(list.files(corpus_data_targetdir, full.names = TRUE))
         
-        .message(
-          sprintf(
-            "path includes non-ASCII characters, moving binary corpus data to temporary data directory (%s/%s)",
-            data_dir(), tolower(corpus)
-          ),
-          verbose = verbose
+        if (verbose) cli_alert_info(
+          paste0(
+            "path includes non-ASCII characters, moving binary corpus data to ",
+            "temporary data directory ",
+            "{.path {path(data_dir(), tolower(corpus))}}"
+          )
         )
         for (x in list.files(corpus_data_srcdir, full.names = TRUE))
           file.copy(from = x, to = path(corpus_data_targetdir, basename(x)))
@@ -100,9 +116,17 @@ use <- function(pkg, corpus, lib.loc = .libPaths(), tmp = FALSE, verbose = TRUE)
       registry_new = registry(),
       home_dir_new = corpus_data_targetdir
     )
-    cl_load_corpus(corpus = toupper(corpus), registry = registry())
-    cqp_load_corpus(corpus = toupper(corpus), registry = registry())
-    
+    success <- cl_load_corpus(corpus = toupper(corpus), registry = registry())
+    if (isFALSE(success)){
+      if (verbose) cli_alert_danger("failed to load corpus")
+      return(FALSE)
+    }
+    success <- cqp_load_corpus(corpus = toupper(corpus), registry = registry())
+    if (isFALSE(success)){
+      if (verbose) cli_alert_danger("failed to load corpus")
+      return(FALSE)
+    }
+
     properties <- sapply(
       corpus_properties(corpus, registry = registry()),
       function(p)
@@ -118,8 +142,10 @@ use <- function(pkg, corpus, lib.loc = .libPaths(), tmp = FALSE, verbose = TRUE)
     if (nchar(additional_info) > 0L)
       additional_info <- sprintf(" (%s)", additional_info)
     
-    .message(sprintf("corpus loaded: %s%s", toupper(corpus), additional_info), verbose = verbose)
+    if (verbose) cli_alert_success(
+      sprintf("corpus loaded: %s%s", toupper(corpus), additional_info)
+    )
   }
   
-  invisible(NULL)
+  invisible(TRUE)
 }